# openapi/aca-openapi-v2.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "ACA-Upload-API-v2",
    "description" : "Microservice to manage PagoPA GPD Upload",
    "termsOfService" : "https://www.pagopa.gov.it/",
    "version" : "0.5.2"
  },
  "servers" : [ {
    "url" : "https://{host}{basePath}",
    "variables" : {
      "basePath" : {
        "default" : "/upload/aca/debt-positions-service/v2",
        "enum" : [ "/upload/aca/debt-positions-service/v2" ]
      },
      "host" : {
        "default" : "api.platform.pagopa.it",
        "enum" : [ "api.uat.platform.pagopa.it", "api.platform.pagopa.it" ]
      }
    }
  } ],
  "paths" : {
...
    "/brokers/{broker-code}/organizations/{organization-fiscal-code}/debtpositions/file/{file-id}/report" : {
      "get" : {
        "tags" : [ "Massive operation observability APIs - v2" ],
        "summary" : "Returns the debt positions upload report.",
        "operationId" : "get-debt-positions-upload-report",
        "parameters" : [ {
          "name" : "broker-code",
          "in" : "path",
          "description" : "The broker code",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        }, {
          "name" : "organization-fiscal-code",
          "in" : "path",
          "description" : "The organization fiscal code",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        }, {
          "name" : "file-id",
          "in" : "path",
          "description" : "The unique identifier for file upload",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Upload report found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadReportDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Malformed request.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "401" : {
            "description" : "Wrong or missing function key.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          },
          "404" : {
            "description" : "Upload report not found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "425" : {
            "description" : "The report has not been generated yet, wait for the operations to complete.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests.",
            "content" : {
              "text/json" : { }
            }
          },
          "500" : {
            "description" : "Service unavailable.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ]
      }
    },
    "/brokers/{broker-code}/organizations/{organization-fiscal-code}/debtpositions/file/{file-id}/status" : {
      "get" : {
        "tags" : [ "Massive operation observability APIs - v2" ],
        "summary" : "Returns the debt positions upload status.",
        "operationId" : "get-debt-positions-upload-status",
        "parameters" : [ {
          "name" : "broker-code",
          "in" : "path",
          "description" : "The broker code",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        }, {
          "name" : "organization-fiscal-code",
          "in" : "path",
          "description" : "The organization fiscal code",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        }, {
          "name" : "file-id",
          "in" : "path",
          "description" : "The unique identifier for file upload",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Upload found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadStatusDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Malformed request.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "401" : {
            "description" : "Wrong or missing function key.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          },
          "404" : {
            "description" : "Upload not found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests.",
            "content" : {
              "text/json" : { }
            }
          },
          "500" : {
            "description" : "Service unavailable.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ]
      }
    },
    "/brokers/{broker-code}/organizations/{organization-fiscal-code}/debtpositions/files" : {
      "get" : {
        "tags" : [ "Massive operation observability APIs - v2" ],
        "summary" : "Returns the list of fileIds for a broker/organization in the given date range (max 7 days).",
        "operationId" : "get-debt-positions-fileids",
        "parameters" : [ {
          "name" : "broker-code",
          "in" : "path",
          "description" : "The broker code",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        }, {
          "name" : "organization-fiscal-code",
          "in" : "path",
          "description" : "The organization fiscal code",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "description" : "Start date (YYYY-MM-DD), Europe/Rome",
          "required" : false,
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example" : "2025-09-01"
        }, {
          "name" : "to",
          "in" : "query",
          "description" : "End date (YYYY-MM-DD), Europe/Rome",
          "required" : false,
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example" : "2025-09-06"
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Max items per page (default 100, min 100, max 500)",
          "required" : false,
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 100
          }
        }, {
          "name" : "x-continuation-token",
          "in" : "header",
          "description" : "Continuation token (opaque). Pass it back to get the next page.",
          "schema" : {
            "type" : "string",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "FileIds retrieved.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileIdListResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "401" : {
            "description" : "Wrong or missing function key.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          },
          "410" : {
            "description" : "Gone",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests.",
            "content" : {
              "text/json" : { }
            }
          },
          "500" : {
            "description" : "Service unavailable.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "FileIdListResponse" : {
        "type" : "object",
        "properties" : {
          "fileIds" : {
            "type" : "array",
            "description" : "List of file identifiers",
            "items" : {
              "type" : "string"
            }
          },
          "size" : {
            "type" : "integer",
            "description" : "Number of items in this page",
            "format" : "int32"
          },
          "hasMore" : {
            "type" : "boolean",
            "description" : "True if another page is available (x-continuation-token present)"
          }
        }
      },
      "OperationStatus" : {
        "type" : "string",
        "enum" : [ "IN_PROGRESS", "COMPLETED", "COMPLETED_WITH_WARNINGS", "COMPLETED_UNSUCCESSFULLY" ]
      },
      "ProblemJson" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "description" : "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable"
          },
          "status" : {
            "type" : "integer",
            "description" : "The HTTP status code generated by the origin server for this occurrence of the problem.",
            "format" : "int32",
            "example" : 400
          },
          "detail" : {
            "type" : "string",
            "description" : "A human readable explanation specific to this occurrence of the problem.",
            "example" : "There was an error processing the request"
          }
        },
        "description" : "Object returned as response in case of an error."
      },
      "ResponseEntryDTO" : {
        "type" : "object",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 400
          },
          "statusMessage" : {
            "type" : "string",
            "example" : "Bad request caused by invalid email address"
          },
          "iupds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "UploadReportDTO" : {
        "type" : "object",
        "properties" : {
          "fileId" : {
            "type" : "string"
          },
          "processedItem" : {
            "type" : "integer",
            "format" : "int32"
          },
          "submittedItem" : {
            "type" : "integer",
            "format" : "int32"
          },
          "responses" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ResponseEntryDTO"
            }
          },
          "startTime" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2024-10-08T14:55:16.302Z"
          },
          "endTime" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "UploadStatusDTO" : {
        "type" : "object",
        "properties" : {
          "fileId" : {
            "type" : "string"
          },
          "processedItem" : {
            "type" : "integer",
            "format" : "int32"
          },
          "submittedItem" : {
            "type" : "integer",
            "format" : "int32"
          },
          "startTime" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2024-10-08T14:55:16.302Z"
          },
          "operationStatus" : {
            "$ref" : "#/components/schemas/OperationStatus"
          }
        }
      }
    },
    "securitySchemes" : {
      "ApiKey" : {
        "type" : "apiKey",
        "name" : "ApiKey",
        "in" : "header"
      }
    }
  }
}